:recycle: Remove Wallet module and replace by AccountDetails component
