+++ src/assets/svgs/CompletedIllustrationSvg.js
@@ -7,7 +7,7 @@ import { themes, colors } from 'constants/styleGuide';
 export default function CompletedIllustrationSvg({ height = 97, width = 59, style }) {
   const { theme } = useTheme();
 
-  const baseColor = theme === themes.light ? colors.light.zodiacBlue : colors.dark.white;
+  const color = theme === themes.light ? colors.light.zodiacBlue : colors.dark.white;
 
   return (
     <Svg width={width} height={height} viewBox="0 0 59 97" fill="none" style={style}>
@@ -19,14 +19,14 @@ export default function CompletedIllustrationSvg({ height = 97, width = 59, styl
         d="M57.2394 41.9316C55.1973 38.2096 50.7419 37.2232 46.7228 39.4192C37.914 44.2672 35.3707 53.4514 35.5193 53.8236C35.6678 54.1958 44.8014 56.9687 53.6658 52.158C57.6385 49.9806 59.0958 45.3838 57.2394 41.9316Z"
         fill="#2BD67B"
       />
-      <Path d="M29.189 5.67969H28.2612V81.2188H29.189V5.67969Z" fill={baseColor} />
+      <Path d="M29.189 5.67969H28.2612V81.2188H29.189V5.67969Z" fill={color} />
       <Path
         d="M43.4334 49.3682L28.5054 56.4453L28.9022 57.2866L43.8302 50.2094L43.4334 49.3682Z"
-        fill={baseColor}
+        fill={color}
       />
       <Path
         d="M19.0594 24.3442L18.4028 25.002L28.3959 35.0268L29.0525 34.369L19.0594 24.3442Z"
-        fill={baseColor}
+        fill={color}
       />
       <Path
         d="M39.9372 1.04554C39.6444 0.72411 39.2895 0.465737 38.894 0.286113C38.4985 0.106488 38.0707 0.00934604 37.6366 0.000641802C37.2025 -0.00806243 36.7712 0.07185 36.3688 0.235476C35.9665 0.399103 35.6015 0.643039 35.2962 0.952476C34.283 1.84376 33.4605 2.93159 32.8782 4.15033C32.2959 5.36908 31.9659 6.69347 31.9082 8.04358C33.2493 8.18132 34.6044 8.0418 35.8896 7.63364C37.1748 7.22547 38.3629 6.55737 39.3803 5.67058C40.0659 5.13128 40.5097 4.34104 40.6141 3.47369C40.7186 2.60634 40.4751 1.73292 39.9372 1.04554Z"
@@ -34,11 +34,11 @@ export default function CompletedIllustrationSvg({ height = 97, width = 59, styl
       />
       <Path
         d="M34.7663 4.6026L28.3535 11.0312L29.0099 11.6892L35.4226 5.26059L34.7663 4.6026Z"
-        fill={baseColor}
+        fill={color}
       />
       <Path
         d="M55.5132 97.0001H0.962402L1.37082 96.3023C4.09569 91.5741 8.01374 87.6481 12.7314 84.9185C17.449 82.1889 22.8001 80.752 28.2471 80.752C33.6941 80.752 39.0452 82.1889 43.7628 84.9185C48.4804 87.6481 52.3985 91.5741 55.1234 96.3023L55.5132 97.0001ZM2.60532 96.0696H53.9167C51.2228 91.6741 47.4501 88.044 42.9587 85.5259C38.4673 83.0078 33.4071 81.6856 28.261 81.6856C23.1149 81.6856 18.0547 83.0078 13.5633 85.5259C9.07191 88.044 5.29923 91.6741 2.60532 96.0696Z"
-        fill={baseColor}
+        fill={color}
       />
     </Svg>
   );
