--- src/modules/Accounts/components/Footer.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import { View } from 'react-native';
-import withTheme from 'components/shared/withTheme';
-import getStyles from './styles';
-
-const Footer = ({ styles }) => <View style={styles.footer}></View>;
-
-export default withTheme(Footer, getStyles());
