--- src/modules/Accounts/components/index.js
@@ -1,3 +1,0 @@
-export { default as Manager } from './Manager';
-export { default as LoadingState } from './LoadingState';
-export { default as LockedBalanceDetails } from './LockedBalanceDetails';
