+++ src/modules/Transactions/hooks/useTransactionAssets.js
@@ -1,20 +1,9 @@
 /* eslint-disable complexity */
-import { themes } from 'constants/styleGuide';
+import React from 'react';
+
 import { useTheme } from 'contexts/ThemeContext';
 import { useCurrentAccount } from 'modules/Accounts/hooks/useCurrentAccount';
-
-import setSecondPassphraseLight from 'assets/images/txDetail/second-passphrase-light.png';
-import setSecondPassphraseDark from 'assets/images/txDetail/second-passphrase-dark.png';
-import registerDelegateLight from 'assets/images/txDetail/delegate-registration-light.png';
-import registerDelegateDark from 'assets/images/txDetail/delegate-registration-dark.png';
-import voteLight from 'assets/images/txDetail/vote-light.png';
-import voteDark from 'assets/images/txDetail/vote-dark.png';
-import transferDark from 'assets/images/txDetail/transfer-dark.png';
-import transferLight from 'assets/images/txDetail/transfer-light.png';
-import txUnlockLight from 'assets/images/txDetail/tx-unlock.png';
-import txUnlockDark from 'assets/images/txDetail/tx-unlock-dark.png';
-import txUnknownLight from 'assets/images/txDetail/tx-unknown-light.png';
-import txUnknownDark from 'assets/images/txDetail/tx-unknown-dark.png';
+import TransactionSvg from 'assets/svgs/TransactionSvg';
 import { MODULE_COMMAND_NAMES } from '../constants';
 import getTransactionRowStyles from '../components/TransactionRow/styles';
 
@@ -24,14 +13,16 @@ import getTransactionRowStyles from '../components/TransactionRow/styles';
  * @param {Object} transaction - The transaction object to get the assets from.
  * @returns {Object} - The assets of the transaction.
  */
-export function useTransactionAssets(transaction) {
+export function useTransactionAssets({ transaction, style }) {
   const [currentAccount] = useCurrentAccount();
 
-  const { styles, theme } = useTheme({
+  const { styles } = useTheme({
     styles: getTransactionRowStyles(),
   });
 
-  let assets = {};
+  let assets = {
+    icon: <TransactionSvg moduleCommand={transaction.moduleCommand} style={style?.icon} />,
+  };
 
   if (transaction.moduleCommand === MODULE_COMMAND_NAMES.tokenTransfer) {
     if (currentAccount.metadata.address !== transaction.sender.address) {
@@ -59,7 +50,6 @@ export function useTransactionAssets(transaction) {
         ...assets,
         type: 'tokenTransfer',
         title: 'Token Transfer',
-        image: theme === themes.light ? transferLight : transferDark,
       };
       break;
 
@@ -68,7 +58,6 @@ export function useTransactionAssets(transaction) {
         ...assets,
         type: 'tokenCrossChaintransfer',
         title: 'Cross-chain Transfer',
-        image: theme === themes.light ? transferLight : transferDark,
       };
       break;
 
@@ -77,7 +66,6 @@ export function useTransactionAssets(transaction) {
         ...assets,
         type: 'registerMultisignatureGroup',
         title: 'Register multisignature group',
-        image: theme === themes.light ? setSecondPassphraseLight : setSecondPassphraseDark,
       };
       break;
 
@@ -86,7 +74,6 @@ export function useTransactionAssets(transaction) {
         ...assets,
         type: 'registerDelegate',
         title: 'Delegate registration',
-        image: theme === themes.light ? registerDelegateLight : registerDelegateDark,
       };
       break;
 
@@ -95,8 +82,6 @@ export function useTransactionAssets(transaction) {
         ...assets,
         type: 'reportDelegateMisbehavior',
         title: 'Report Delegate Misbehavior',
-        // TODO: Add custom image.
-        image: theme === themes.light ? txUnknownLight : txUnknownDark,
       };
       break;
 
@@ -105,7 +90,6 @@ export function useTransactionAssets(transaction) {
         ...assets,
         type: 'unlockToken',
         title: 'Unlock',
-        image: theme === themes.light ? txUnlockLight : txUnlockDark,
       };
       break;
 
@@ -114,8 +98,6 @@ export function useTransactionAssets(transaction) {
         ...assets,
         type: 'updateGeneratorKey',
         title: 'Update Generator Key',
-        // TODO: Add custom image.
-        image: theme === themes.light ? txUnknownLight : txUnknownDark,
       };
       break;
 
@@ -124,7 +106,6 @@ export function useTransactionAssets(transaction) {
         ...assets,
         type: 'voteDelegate',
         title: 'Vote',
-        image: theme === themes.light ? voteLight : voteDark,
       };
       break;
 
@@ -133,8 +114,6 @@ export function useTransactionAssets(transaction) {
         ...assets,
         type: 'reclaimLSK',
         title: 'Reclaim LSK',
-        // TODO: Add custom image.
-        image: theme === themes.light ? txUnknownLight : txUnknownDark,
       };
       break;
 
@@ -143,8 +122,6 @@ export function useTransactionAssets(transaction) {
         ...assets,
         type: 'registerkeys',
         title: 'Register keys',
-        // TODO: Add custom image.
-        image: theme === themes.light ? registerDelegateLight : registerDelegateDark,
       };
       break;
