--- src/modules/Wallet/components/index.js
@@ -1,1 +1,0 @@
-export { default as AccountSummary } from './AccountSummary';
