--- src/modules/Wallet/utils/index.js
@@ -1,1 +1,0 @@
-export * from './account';
